+++ examples/waku-chat/src/chat.rs
@@ -1,9 +1,9 @@
 use crate::network::*;
 use async_trait::async_trait;
-use overwatch::services::handle::ServiceStateHandle;
-use overwatch::services::relay::{NoMessage, OutboundRelay};
-use overwatch::services::state::{NoOperator, NoState};
-use overwatch::services::{ServiceCore, ServiceData, ServiceId};
+use overwatch_rs::services::handle::ServiceStateHandle;
+use overwatch_rs::services::relay::{NoMessage, OutboundRelay};
+use overwatch_rs::services::state::{NoOperator, NoState};
+use overwatch_rs::services::{ServiceCore, ServiceData, ServiceId};
 use serde::{Deserialize, Serialize};
 use tokio::sync::mpsc::channel;
