+++ overwatch-derive/Cargo.toml
@@ -5,6 +5,17 @@ edition = "2021"
 authors = [
     "Daniel Sanchez Quiros <danielsq@status.im>"
 ]
+license-file = "LICENSE"
+homepage = "https://github.com/logos-co/Overwatch"
+repository = "https://github.com/logos-co/Overwatch"
+description = "Overwatch derive macros"
+readme = "README.md"
+keywords = ["async", "services"]
+categories = ["Asynchronous"]
+exclude = [
+    "./tests",
+    ".github",
+]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [lib]
