+++ overwatch-rs/src/utils/const_checks.rs
